First Steps
Eliminated the shuttle_master from the gameticker, it gave a runtime
error that forbid the Join Game window from appearing

Commented out many of the jobs and added the new ones.

Added some new floors and the ability to change day and night times for
the admins.

Added teleports that lead players to either the medieval or the
omegacorp areas, they don't check validicity, though.

Changed preferences so the underwear only affects people of the
Omegacorp.

Eliminated many of the species.