It's only been a week of coding, but there's changes
Changed the name of the job lists to better organise the code.

Changed a variable in the wizard/artefact.dm code, that variable should
be using the "all_species" list, but instead it uses it's own
"all_species", that's pretty mysterious

Added the jobs

Eliminated the special organs of the Paradise species which aren't the
ones I left off in the code, basically anything alien or robotic is out.

Added the Omegacorp IDs

**MASSIVE SHIT** C.M.A. Admin has stopped working, it's probably because
of the addition of these two new species I made, I've been checking
around and there was a log file where a "null" species was loaded in the
.tmpl file for the C.M.A. window, that's what's giving the error, no
doubts, I'll probably have to figure that out later.

Added the teleport_fancy to go to the medieval world through the
"Join_Late" Hub, I'll turn that hub into a "new players" area later, for
now, it's just there for show.

Commented ALL THE LOADOUT THINGS until I make the new ones.

Changed the default UI_Style to Slimecore, because green is more
fantastical

REMEMBER TO MAKE CLOTHING ADJUSTMENTS FOR DWARVES AND KOBOLDS.

Created the "clothing/smallsize" object, the idea is to make clothing
that only Kobolds and Dwarves can wear, whereas they can actually wear
big people clothing! CHEATERS!

Added some clothing for the omegacorp folk

Eliminated even more shuttle stuff

Added many omegacorp-only items, they are on the omegacorp folder, of
course, under modules.

ICOOOOOOOOONS

SOUUUUUUUUUNDS

MAAAAAAAAAAAAP