Almost done with basic rules and Omegacorp
Created lists for the medieval areas, it'll help with the day and night
transition in a later time.

Now round end comes in a Restart Vote called via configuration file
altering.

Attempting to organise omegacorp clothing.

FUCKING PODDOORS

I've added the "is_disguised" var to an item in order to prevent the
item from updating icons while it is disguised, although I haven't
updated all the update_icon methods to allow for that -.-

A new tile type, worker_wood_mini

NOW WE HAVE MANUALS WITH THE BACKSTORY AND HOW THE GAME SHOULD BE
PLAYED!! I'll complete them as time passes.

Metal shelves added

Now floors will check if they are in a medieval area or if they are in a
Omegacorp area, if they are in a medieval area, they'll change the turf
to GRASS after being destroyed (This must be updated later for desert
and snow maps)

Admins cannot use the Shuttle Verbs (They can still use the Secret
Shuttle Verbs tho)

The teleporters will prevent people from smuggling manuals into reality

ARMOR REFORGER FOR OMEGACORP, it reinforces and changes some armors
sprites, the reinforce is pretty minor, but it's pretty cheap too.

A HEAT CONTROLLER to control Omegacorp from being too fucking OP,
anytime they use a computer, this thing drains energy, it still needs a
way to be cooled down, tho.

Some updates for the disguiser.

MAINFRAMES, they are neccesary for the Phonebooth to work at it's
fullest potential. They are easy to set up, but the Power Mainframe will
explode as commanded by the heat controller.

UPDATES TO THE PHONEBOOTH

The phonebooth controller is now a sonic screwdriver.

MANY FIXES FOR ALL THE COMPUTERS.

THE POWER OF CSS!!!!!

NEW SOUNDS!!! (Freesound.org is a marvelous webpage)