Changes up to 13/03
Added a sea area

Changing some floor stuff, now it regains the bug that makes it so
having an explosion take place in the omegacorp base creates grass, fix
this shit later

Some changes to the floor and creating dirt floor, which is basically
floor that won't create grass upon destruction, it has a bug with a
runtime in a process, bad stuff, the idea is that it can be changed into
grass by using seeds

Moved water from misc floor to fancy floor

Disabled the automated fungus + some flowers weed invasions while we
control the plant types we want to grow

Added SOME fixes to the caravan

Added mine dangers file which will contain natural dangers that can
happen in the mine, monsters should be in a separate file

Improved the mine generation

ATTEMPTED to fix the money values, but we'll probably just make it a
plain sprite from now on

Added woodworks, changed trees, good stuff

The saloon door is now kind of a sprite that has an animation when you
cross it, it will also stop animals from crossing it when it gets
covered with barbed wire (Locked?)

ADDED THE SHIP, FIX MORE SHIT LATER

Separated the farm animals into files as we want them to all have their
own life patterns

Added a kobold smuggler to be added in the store's basement.

Added some changes to the way humans breathe so they can breathe inside
SOME things that they shouldn't be able.

Added a mana bar on the right side of the screen.

other stuff